requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
google-api-python-client==2.136.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pytz==2024.2
tzdata==2024.2
uvicorn==0.32.0
fastapi==0.115.5
requests==2.31.0
openai>=1.12.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==6.0.0
